fix: resolve type errors and remove legacy Svelte 4 reactivity workarounds
- Remove unnecessary self-assignment for reactivity (Svelte 5 handles this)
- Add Set<string> generic type in BookmarkList
- Move getFishSpeechModels from CharConfig to TTSSettings
- Disable strict mode in tsconfig temporarily
