+++ src/lib/SideBars/TTS/TTSSettings.svelte
@@ -50,8 +50,34 @@
         }
     })
 
-    function getFishSpeechModels() {
-        throw new Error("Function not implemented.")
+    /**
+     * Fetch model list from Fish Speech API.
+     */
+    async function getFishSpeechModels() {
+        try {
+            const res = await fetch(`https://api.fish.audio/model?self=true`, {
+                headers: {
+                    Authorization: `Bearer ${DBState.db.fishSpeechKey}`,
+                },
+            })
+            const data = await res.json()
+            console.log(data.items)
+            console.log(DBState.currentChar)
+
+            if (Array.isArray(data.items)) {
+                fishSpeechModels = data.items.map((item: { _id?: string; title?: string; description?: string }) => ({
+                    _id: item._id || "",
+                    title: item.title || "",
+                    description: item.description || "",
+                }))
+            } else {
+                console.error("Expected an array of items, but received:", data.items)
+                fishSpeechModels = []
+            }
+        } catch (error) {
+            console.error("Error fetching fish speech models:", error)
+            fishSpeechModels = []
+        }
     }
 </script>
