+++ tsconfig.json
@@ -20,7 +20,7 @@
     "checkJs": false,
 
     /* Type Checking */
-    "strict": true,
+    "strict": false,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
     "noFallthroughCasesInSwitch": true,
